Fixed rating system

+++ api_yamdb/reviews/models.py
@@ -33,7 +33,7 @@ class Title(models.Model):
         related_name="titles")
 
     genre = models.ManyToManyField(Genre, through='GenreTitle')
-    rating = models.FloatField(blank=True, null=True)
+    rating = models.IntegerField(blank=True, null=True)
 
     def __str__(self):
         return self.name
@@ -66,13 +66,6 @@ class Review(models.Model):
         'Дата публикации',
         auto_now_add=True,
     )
-
-#    class Meta:
-#        constraints = [
-#            models.UniqueConstraint(fields=['title', 'author'],
-#                                    name='unique_author_review')
-#        ]
-#        unique_together = ['title', 'author']
 
 
 class Comment(models.Model):
